Reorganize TP sections to clarify steps
In the Ubuntu procedure, the flow for the transaction processor
information isn't clear.  It's hard to tell what's a step and
what's background information or tips for a production environment.

I reorganized the Transaction Processor sections to group
related subsections together and to make each section title
a step (except for the list of supported transaction families).

Also added a new section, Stopping Sawtooth Components.

Signed-off-by: Anne Chenette <chenette@bitwise.io>